requirements.txt
asgiref==3.2.10
Django==3.1.1
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.1
gunicorn==20.0.4
Markdown==3.2.2
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
